Add Basket API endpoints and change price type to decimal
Created new endpoints for deleting, storing, and retrieving baskets in the Basket API. Updated the price type from double to decimal in the Catalog API, ensuring consistency and precision across the product-related operations.

+++ src/Services/Catalog.API/Models/Product.cs
@@ -2,11 +2,12 @@
 {
     public class Product
     {
+        //Mapping
         public Product()
         {
 
         }
-        public Product(Guid _Id, string _Name, List<string> _Category, string _Description, string _ImageUrl, double _Price)
+        public Product(Guid _Id, string _Name, List<string> _Category, string _Description, string _ImageUrl, decimal _Price)
         {
             Id = _Id;
             Name = _Name;
@@ -20,6 +21,6 @@
         public List<string> Category { get; set; } = default!;
         public string Description { get; set; } = string.Empty;
         public string ImageUrl { get; set; } = string.Empty;
-        public double Price { get; set; }
+        public decimal Price { get; set; }
     }
 }
